requirements.txt
climb-astro
numpy
pandas
matplotlib
scikit-learn
seaborn
joblib
pyclustertend